api-gateway/requirements.txt
Flask
requests
PyJWT
Flask-Cors
